% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{addComponent}
\alias{addComponent}
\title{Add a new component to sim file}
\usage{
addComponent(sim, component, parent = "<system name=\\"paddock\\"")
}
\arguments{
\item{sim}{A character vector of sim}

\item{component}{A character vec of new component}

\item{parent}{A character string containing a regular expression 
to match parent component}
}
\description{
Add a new component to sim file
}
