% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{changeAttribute}
\alias{changeAttribute}
\title{Change attribute for a xml element}
\usage{
changeAttribute(xml, attri, new)
}
\arguments{
\item{xml}{A xml element}

\item{attri}{The attribute name}

\item{new}{The new attribute value}
}
\description{
Change attribute for a xml element
}
