% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/componentTemplate.R
\name{componentTemplate}
\alias{componentTemplate}
\title{Get template by component}
\usage{
componentTemplate(component)
}
\arguments{
\item{component}{The component name}
}
\description{
Get template by component
}
