% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim.R
\name{fertiliseAtFixed}
\alias{fertiliseAtFixed}
\title{Add fertilizaton at fixed date component}
\usage{
fertiliseAtFixed(amount, type, date, name = "Fertilise on fixed date")
}
\arguments{
\item{amount}{Irrigation amount}
}
\description{
Add fertilizaton at fixed date component
}

