% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim.R
\name{fertiliseAtSowing}
\alias{fertiliseAtSowing}
\title{Add fertilizaton at sowing component}
\usage{
fertiliseAtSowing(amount, type, name = "Fertilise at sowing")
}
\arguments{
\item{amount}{Irrigation amount}
}
\description{
Add fertilizaton at sowing component
}

