% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{findXmlValue}
\alias{findXmlValue}
\title{Find the values with key from xml}
\usage{
findXmlValue(xml, key)
}
\arguments{
\item{xml}{the xml document}

\item{key}{the name of xml element}
}
\description{
Find the values with key from xml
}
