% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ncAPSIM.R
\name{ncCalculateApplyIndex}
\alias{ncCalculateApplyIndex}
\title{Calculate index for variables of applying a function}
\usage{
ncCalculateApplyIndex(nc, depth = length(nc$dim) - 1)
}
\arguments{
\item{nc}{An object of class ncdf4 (as returned from ncOpen), indicating what file to read from.}

\item{depth}{The depth for variables of applying a function.}
}
\description{
Calculate index for variables of applying a function
}

