% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ncAPSIM.R
\name{ncCreateApsim}
\alias{ncCreateApsim}
\title{Create netcdf file for APSIM}
\usage{
ncCreateApsim(filename, years, factors, traits, is.close = TRUE,
  global.att = NULL, ...)
}
\arguments{
\item{filename}{The filename to save netcdf file.}

\item{years}{A vector for year demension.}

\item{factors}{A list for all dimensions in the netcdf.}

\item{traits}{A vector of traits in the netcdf.}

\item{is.close}{Whether to close netcdf file.}

\item{global.att}{Global attribures}

\item{...}{Attributes of traits, e.g. unit
If false, a nv variable will be returned.}
}
\description{
Create netcdf file for APSIM
}
\examples{
traits <- c('yield', 'flow_das')
years <- 2000:2011
sites <- c('gatton', 'dubbo', 'moree', 'dalby', 'emerald')
date <- format(as.Date(seq(120, 270, by = 5),
    origin = '2010-12-31'), '\\\%d-\\\%b')
cultivar <- c('hartog', 'janz', 'anlace',
    'sunco', 'batavia', 'westonia', 'bolac')
fert_amount_sow <- seq(100, 200, by = 10)
factors <- list(
    site = sites,
    date = date,
    cultivar = cultivar,
    fert_amount_sow = fert_amount_sow)
filename <- 'demo.nc'
ncCreateApsim(filename, years, factors, traits)
# Add attributes to traits
ncCreateApsim(filename, years, factors, traits,
        unit = c('kg/ha', 'das'))
# Add global attributes to nc file
ncCreateApsim(filename, years, factors, traits,
        unit = c('kg/ha', 'das'),
        global.att = list(author = 'Bangyou Zheng'))
}

