% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ncAPSIM.R
\name{ncGetApplyIndex}
\alias{ncGetApplyIndex}
\title{Return index for variables of applying a function}
\usage{
ncGetApplyIndex(nc, ...)
}
\arguments{
\item{nc}{An object of class ncdf4 (as returned from ncOpen), indicating what file to read from.}

\item{...}{Other arguments to specify variables of dimensions}
}
\description{
Return index for variables of applying a function
}

