% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ncAPSIM.R
\name{ncReadApsimOut}
\alias{ncReadApsimOut}
\title{Read APSIM out file}
\usage{
ncReadApsimOut(nc, filenames, start, res.array, yearvar = "Year",
  traits = NULL, ...)
}
\description{
Read APSIM out file
}

