% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsimOut.R
\name{readApsimHead}
\alias{readApsimHead}
\title{Read head information from out file}
\usage{
readApsimHead(filename, headname, split = " *, *", ...)
}
\arguments{
\item{filename}{filename}

\item{headname}{headname}

\item{split}{split}

\item{...}{...}
}
\description{
Read head information from out file
}
