% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim.R
\name{removeComponent}
\alias{removeComponent}
\title{Remove a component from sim file}
\usage{
removeComponent(sim, pattern)
}
\arguments{
\item{sim}{A character vector of sim}

\item{pattern}{A character string containing a regular expression
to match a component}
}
\description{
Remove a component from sim file
}

