% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{replaceElementVlaue}
\alias{replaceElementVlaue}
\title{Replace a row for element name}
\usage{
replaceElementVlaue(value, xml, key = NULL)
}
\arguments{
\item{value}{value}

\item{xml}{xml}

\item{key}{key}
}
\description{
Replace a row for element name
}
