% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/condor.R
\name{simPrefix}
\alias{simPrefix}
\title{Generate the sim prefix with md5}
\usage{
simPrefix(factors)
}
\arguments{
\item{factors}{A data.frame for factors}
}
\description{
Generate the sim prefix with md5
}

