% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{fertiliseAtFixed}
\alias{fertiliseAtFixed}
\title{Add fertilizaton at fixed date component}
\usage{
fertiliseAtFixed(amount, type, date, name = "Fertilise on fixed date")
}
\arguments{
\item{amount}{Irrigation amount}

\item{type}{type}

\item{date}{date}

\item{name}{name}
}
\description{
Add fertilizaton at fixed date component
}
