% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{fertiliseAtSowing}
\alias{fertiliseAtSowing}
\title{Add fertilizaton at sowing component}
\usage{
fertiliseAtSowing(amount, type, name = "Fertilise at sowing")
}
\arguments{
\item{amount}{Irrigation amount}

\item{type}{type}

\item{name}{name}
}
\description{
Add fertilizaton at sowing component
}
