% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim.R
\name{irrigationAtFixedDate}
\alias{irrigationAtFixedDate}
\title{Add irrigation at fixed date component}
\usage{
irrigationAtFixedDate(amount, date, name = "Irrigate on fixed date")
}
\arguments{
\item{amount}{Irrigation amount}
}
\description{
Add irrigation at fixed date component
}

